requirements.txt
asgiref==3.7.2
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.0
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
openpyxl==3.1.5
pillow==11.3.0
pillow_heif==1.1.0
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.3
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.4.4
typing_extensions==4.15.0
tzdata==2023.3
uritemplate==4.2.0
urllib3==2.5.0
